requirements.txt
streamlit==1.36.0
pandas==2.2.2
gspread==6.2.0
google-auth==2.40.0